+++ assets/js/source/kyte-shipyard-site-sidenavitem.js
@@ -50,6 +50,12 @@ $(document).ready(function() {
                         selector:'#Scripts'
                     },
                     {
+                        faicon:'fas fa-book',
+                        label:'Library',
+                        href:'/app/site/?request='+encoded+'#Library',
+                        selector:'#Library'
+                    },
+                    {
                         faicon:'fas fa-photo-video',
                         label:'Media',
                         href:'/app/site/?request='+encoded+'#Media',